Move models into admiral and add tests
