+++ src/admiral/model/__init__.py
@@ -1,0 +1,4 @@
+from .cert import Cert
+from .domain import Domain, Agency
+
+__all__ = ["Cert", "Domain", "Agency"]
